rename opcode metadata sort functions for consistency
there are two sets of sort functions. the first two are used
to sort copies of the metadata tables for output by x86_opcodes,
and to sort metadata tables before expanding their prefixes
during construction of acceleration tables. the second two are
used for sorting and looking up opcodes in acceleration tables.

- x86_opc_data_compare_alpha - use for alphanumeric sort
- x86_opc_data_compare_opcode - used for type and opcode sort
- x86_opc_data_compare_build - used to build acceleration tables
- x86_opc_data_compare_masked - used to lookup acceleration table