move and fix bug in check for ModRM or second opcode byte
given normalization of maps and prefixes, opcodes are one or
two bytes with the second byte being the ModRM byte or the
second byte of a two byte opcode. during decode, this can only
be established after we have matched from one to two opcode
bytes and found a matching encoding record. this sometime
requires putting back one opcode byte. it is better to keep
this together with the logic that reads the advance byte.

in addition to this, the opclen field must be set correctly,
because sometimes the second byte is emitted as part of a
two byte opcode, and sometimes it is emitted as a one byte
opcode with a ModRM byte. make sure to set opclen to one if
the second opcode field is a whole ModRM byte or ModRM
function because the output function writes the ModRM byte
based on the presence of 'x86_enc_f_modrm_{n|r}' flags.