add segment and legacy parsing states to speed up decode
previously the 'top' state parsed segment and prefix bytes as
well as decoding VEX and EVEX instructions which don't have
prefix bytes. this requires an error check in VEX and EVEX
parsing as VEX and EVEX instructions forbid legacy prefixes.

by adding 'segment' and 'legacy' states, it can be assured
that the top state has not parsed a segment or prefix byte
when encountering the start of a VEX or EVEX instruction,
so an error check in VEX and EVEX parsing can be eliminated.

this requires adding two new states:

- 'segment' state is switched to when encountering a segment
  prefix because segment prefixes must come first. 'segment'
  will always parse one segment prefix or terminate the
  instruction as a lone segment prefix with no opcode byte
  when a subsequent segment prefix is encountered which will
  be parsed as a new instruction.

- 'legacy' state is switched to when encountering a regular
  prefix. it parses multiple legacy prefixes but will
  terminate parsing the instruction when encountering an
  invalid VEX or EVEX prefix or segment prefix because they
  must come before other prefixes.

- 'top' remains the only state that processes VEX and EVEX
  prefixes but it branches to 'segment' for segment prefixes
  and 'legacy' for the other legacy prefixes.

'segment' and 'legacy' can now unconditionally terminate the
instruction as lone prefixes if it encounters VEX or EVEX
prefixes, or segment prefixes in the case of the 'legacy'
state which comes after parsing of the segment prefixes.