x86_fuzzer: first cut of our combinatorial disassembly fuzzer
this fuzzer is a simple combinatorial machine code emitter with
subexpression constraints. columns are added to the generator
with predicates to alter the domain of dependent variables.

for example, memory operand SIB columns activate if `rm == sp`
but are otherwise inactive. here is the generator function:

  int x86_gen_next(x86_operand_gen *gen, size_t count)
  {
      for (size_t i = count-1; i < count; i--) {
          if (gen[i].pred && gen[i].pred(gen, i) == 0) continue;
          gen[i].value += gen[i].step;
          if (gen[i].value <= gen[i].end) {
              return 1;
          } else {
              gen[i].value = gen[i].start;
          }
      }
      return 0;
  }

it is simple but powerful generator as predicate functions are
called when evaluating a new column, and these allow altering
the domain of a finite domain variable based on an epxression.