fix opcode index lookup failure for modrm byte clashes
entries with mod != 0b11 constraints (memory operand) result
in multiple entries for different values of mod. these entries
can clash with entries for the same opcodes that have ModRM /n
functions which can cause lookup failures.

find opcodes that have both ModRM /n functions and reg or mem
constraints and create an index of these clashes and use that
to expand the mod field for all entries of the specific opcode.