add mode-dependent address sized register operand `ra`
the `ra` register specifier allows us to specify register
operands for instructions with fixed encodings that have
mode-dependent register size based on the address size.
this has different behaviour to the `.ww` synthesis suffix
because the instructions that use the `ra` operand do not
respond to address or operand size prefixes.