x86_opc_data: eliminate bitfield due to performance issues
GCC is making poor choices for bitfield access sizes when
translating bitfield accesses. the C specification states
that field accesses may use the smallest size possible
although it usually makes more sense to use the size of
the holding type, otherwise it is a waste of time specifying
types on bitfields. the use of irregular access sizes
results in store-to-load stalls. i.e. the compiler stores
a byte to part of an int sized field then later loads the
data back as an integer. that causes a store-to-load stall.

x86_opc_data grows from 12-bytes to 16-bytes although the
larger cache footprint is offset by less cache stalls.